requirements.txt
fastapi
cloudinary
supabase
python-dotenv